fix(delete sandbox): fixed a bug where sepes tried to find resources for a deleted study
Closes #298
